+++ src/modules/featured_releases/featured_releasesType.ts
@@ -1,5 +1,5 @@
 import { selectorRequiredSchema } from '@/core/types/selectorType';
-import { videoFile } from '@/core/utils/helper';
+import { videoFile } from '@/core/utils/formUtlis';
 import { z } from 'zod';
 import { ReleasesType } from '../releases/releasesType';
